+++ lib/src/infrastructure/internals/packets/listeners/typing_packet.dart
@@ -17,7 +17,7 @@ final class TypingPacket implements ListenablePacket {
       serverId: Snowflake.nullable(payload['guild_id']),
       channelId: Snowflake.parse(payload['channel_id']),
       userId: Snowflake.parse(payload['user_id']),
-      timestamp: DateTime.fromMillisecondsSinceEpoch(payload['timestamp']),
+      timestamp: DateTime.fromMillisecondsSinceEpoch(payload['timestamp'] as int),
     );
 
     dispatch(event: Event.typing, params: [typing]);
